# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Legend Academy API Documentation",
        "description": "Legend Pay API description",
        "contact": {
            "email": "o.oyedele@suburbanfiberco.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/modules": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "summary": "Get list of module",
                "description": "Returns list of modules",
                "operationId": "getModulesList",
                "parameters": [
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Modules Retrieved Successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Module"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Modules"
                ],
                "summary": "Store new course module",
                "description": "Returns course module data",
                "operationId": "storeModule",
                "parameters": [
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreModuleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Module created successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Module"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/modules/{id}": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "summary": "Get Module information",
                "description": "Returns modules data",
                "operationId": "getModuleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Module id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Module Retrieved Successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Module"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Modules"
                ],
                "summary": "Update existing module",
                "description": "Returns updated module data",
                "operationId": "updateModule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Module id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateModuleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Module Updated Successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Module"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Modules"
                ],
                "summary": "Delete existing module",
                "description": "Deletes a record and returns no success message",
                "operationId": "deleteModule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Module id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Module Deleted Successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Module": {
                "title": "Module",
                "description": "Module model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Name",
                        "description": "Title of the new course module",
                        "type": "string",
                        "example": "Introduction to workpkace"
                    },
                    "course_id": {
                        "title": "Course ID",
                        "description": "Course id of the new course module",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "pages": {
                        "title": "Page",
                        "description": "Page model",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Page"
                        }
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Module"
                }
            },
            "Page": {
                "title": "Page",
                "description": "Page model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title of the new course module page",
                        "type": "string",
                        "example": "Introduction to workpkace"
                    },
                    "content": {
                        "title": "Content",
                        "description": "Content of the new course module page",
                        "type": "string",
                        "example": "Introduction to workpkace"
                    },
                    "course_id": {
                        "title": "Course ID",
                        "description": "Course id of the new course module page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "module_id": {
                        "title": "Module ID",
                        "description": "Module id of the new course module page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Page Model"
                }
            },
...
            "StoreModuleRequest": {
                "title": "Store Course Module request",
                "description": "Store Course Module request body data",
                "properties": {
                    "title": {
                        "title": "Name",
                        "description": "Title of the new course module",
                        "type": "string",
                        "example": "Introduction to workpkace"
                    },
                    "course_id": {
                        "title": "Course ID",
                        "description": "Course id of the new course module",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "UpdateModuleRequest": {
                "title": "Store Course Module request",
                "description": "Store Course Module request body data",
                "properties": {
                    "title": {
                        "title": "Name",
                        "description": "Title of the new course module",
                        "type": "string",
                        "example": "Introduction to workpkace"
                    },
                    "course_id": {
                        "title": "Course ID",
                        "description": "Course id of the new course module",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    }
}